ib-tb: support magic keyboard backlight
MacBookPro16,1/2/3/4 and MacBookAir9,1 have the magic keyboard, with the
keyboard backlight controlled via the Touch Bar Backlight USB device.

We check that the internal keyboard's USB product ID matches that of
models with the magic keyboard, and then we create a sysfs led class
dev.

When setting the brightness to 15 (one less than the max of 16), it
seems to have about a 50% chance of setting it to dim instead. This
doesn't happen for other values.

Signed-off-by: Orlando Chamberlain <redecorating@protonmail.com>